+++ ext/intl/normalizer/normalizer_normalize.c
@@ -175,7 +175,7 @@ PHP_FUNCTION( normalizer_is_normalized )
 	char*	 	input = NULL;
 	/* form is optional, defaults to FORM_C */
 	zend_long		form = NORMALIZER_DEFAULT;
-	int		input_len = 0;
+	size_t		input_len = 0;
 
 	UChar*	 	uinput = NULL;
 	int		uinput_len = 0;
